[Add]( Mongo + Dynamo ) Interface 작업 완료

+++ basic_crud/graphql/dynamo_movie.js
@@ -1,29 +1,52 @@
 import Movie from "../models/dynamo_movies";
 
-async function dynamo_getMovies() {
+async function getMovie(
+  title = null,
+  curpage = null,
+  perpage = null,
+  orderby = null
+) {
   try {
-    const result = await Movie.scan().exec();
-    console.log(result);
-    return result;
+    let movies;
+    if (title) {
+      movies = await Movie.scan({ title: title }).exec();
+    } else if (orderby) {
+      let key = Object.getOwnPropertyNames(orderby);
+      let order = orderby[key[0]];
+      console.log(key, order);
+
+      // todo
+      if (order == "asc") {
+        movies = await Movie.scan().exec(function (err, data) {
+          if (err) {
+            console.log(err);
+            throw err;
+          }
+          data.sort();
+          console.log(data);
+          return data;
+        });
+        // movies = await Movie.query("title").eq(key[0]).sort("ascending").exec();
+        // console.log(movies);
+      } else if (order == "desc") {
+        // movies = await Movie.scan().exec().sort("descending").exec();
+      } else {
+        movies = null;
+      }
+    } else if (curpage != null && perpage != null) {
+      //   todo
+
+      movies = await Movie.scan().limit(perpage).exec();
+    } else {
+      movies = await Movie.scan().exec();
+    }
+    return movies;
   } catch (err) {
     console.log(err);
-    throw err;
   }
 }
 
-async function dynamo_findMovie(title) {
-  try {
-    // const result = await Movie.query("title").eq(title).exec();
-    const result = await Movie.scan({ title: title }).exec();
-    console.log(result);
-    return result;
-  } catch (err) {
-    console.log(err);
-    throw err;
-  }
-}
-
-async function dynamo_addMovie(title, score) {
+async function addMovie(title, score) {
   const result = new Movie({
     title: title,
     score: score,
@@ -37,24 +60,31 @@ async function dynamo_addMovie(title, score) {
   }
 }
 
-async function dynamo_deleteMovie(title) {
+async function deleteMovie(title) {
   try {
     const result = await Movie.get({ title: title });
     console.log(result);
-    result.delete();
-    return true;
+    if (result != null) {
+      result.delete();
+      return true;
+    }
   } catch (err) {
     console.log(err);
     return false;
   }
 }
 
-async function dynamo_updateMovie(title, score) {
+// title 없는 내용 update 시 예외처리 추가할것.
+async function updateMovie(title, score) {
   try {
-    const result = await Movie.get({ title: title });
-    result.score = score;
-    await result.save();
-    return result;
+    const exist = await Movie.get({ title: title });
+
+    if (exist != null) {
+      const result = await Movie.update({ title: title }, { score: score });
+      return result;
+    } else {
+      return null;
+    }
   } catch (err) {
     console.log(err);
     throw err;
@@ -62,9 +92,8 @@ async function dynamo_updateMovie(title, score) {
 }
 
 module.exports = {
-  dynamo_getMovies,
-  dynamo_addMovie,
-  dynamo_findMovie,
-  dynamo_deleteMovie,
-  dynamo_updateMovie,
+  getMovie,
+  addMovie,
+  deleteMovie,
+  updateMovie,
 };
